+++ src/query-builder/transformer/DocumentToEntityTransformer.ts
@@ -96,6 +96,9 @@ export class DocumentToEntityTransformer {
                     });
 
                 } else {
+                    if (embedded.embeddeds.length && !entity[embedded.propertyName]) 
+                        entity[embedded.propertyName] = embedded.create();
+                    
                     embedded.columns.forEach(column => {
                         const value = document[embedded.prefix][column.databaseNameWithoutPrefixes];
                         if (value === undefined) return;